LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/stretch/stretch-alias-max-block-size-001.tentative-expected.txt

CONSOLE MESSAGE: No valid data-* attributes found in selector list : [data-expected-height]:not([skip-second-pass])








FAIL [data-expected-height] 1 assert_equals:
<div class="test" data-expected-height="45"></div>
height expected 45 but got 50
FAIL [data-expected-height] 2 assert_equals:
<canvas class="test" data-expected-height="45"></canvas>
height expected 45 but got 50
FAIL [data-expected-height] 3 assert_equals:
<input class="test" data-expected-height="45">
height expected 45 but got 50
FAIL [data-expected-height] 4 assert_equals:
<textarea class="test" data-expected-height="45"></textarea>
height expected 45 but got 50
FAIL [data-expected-height] 5 assert_equals:
<button class="test" data-expected-height="45"></button>
height expected 45 but got 40
FAIL [data-expected-height] 6 assert_equals:
<div class="test" style="float: left" data-expected-height="45"></div>
height expected 45 but got 50
FAIL [data-expected-height] 7 assert_equals:
<div class="test" style="position: absolute; inset-block: 0;" data-expected-height="55"></div>
height expected 55 but got 50
PASS [data-expected-height] 8
FAIL [data-expected-height] 9 assert_equals:
<div class="test" style="position: absolute; inset-block-start: 10px" data-expected-height="45"></div>
height expected 45 but got 50
FAIL [data-expected-height] 10 assert_equals:
<div class="test" style="position: absolute; inset-block-end: 10px" data-expected-height="45"></div>
height expected 45 but got 50
FAIL [data-expected-height] 11 assert_equals:
<div class="test" style="position: absolute; inset-block-end: 55px" data-expected-height="10"></div>
height expected 10 but got 50
FAIL [data-expected-height] 12 assert_equals:
<div class="test" style="position: absolute; inset-block-start: 55px" data-expected-height="10"></div>
height expected 10 but got 50
FAIL [data-expected-height] 13 assert_equals:
<div class="test" style="position: absolute; inset-block: 0;" data-expected-height="55"></div>
height expected 55 but got 50
FAIL [data-expected-height] 14 assert_equals:
<div class="test" style="position: absolute" data-expected-height="55"></div>
height expected 55 but got 50
FAIL [data-expected-height] 15 assert_equals:
<div class="test" style="position: absolute; inset-block-start: 10px" data-expected-height="45"></div>
height expected 45 but got 50
FAIL [data-expected-height] 16 assert_equals:
<div class="test" style="position: absolute; inset-block-end: 10px" data-expected-height="45"></div>
height expected 45 but got 50
FAIL [data-expected-height] 17 assert_equals:
<div class="test" style="position: absolute; inset-block-end: 55px" data-expected-height="10"></div>
height expected 10 but got 50
FAIL [data-expected-height] 18 assert_equals:
<div class="test" style="position: absolute; inset-block-start: 55px" data-expected-height="10"></div>
height expected 10 but got 50
FAIL [data-expected-height] 19 assert_equals:
<div class="test" data-expected-height="45"></div>
height expected 45 but got 50
PASS [data-expected-height] 20
FAIL [data-expected-height] 21 assert_equals:
<div class="test" data-expected-height="55"></div>
height expected 55 but got 50
FAIL [data-expected-height] 22 assert_equals:
<div class="test" data-expected-height="30"></div>
height expected 30 but got 50
FAIL [data-expected-height] 23 assert_equals:
<div class="test" data-expected-height="65" skip-second-pass=""></div>
height expected 65 but got 50
